wada/prompts/debate.py
from typing import Any

from wada.prompts import TextPrompt, TextPromptDict
from wada.typing import RoleType


class DebatePromptTemplateDict(TextPromptDict):

    DEBATER_PROMPT = TextPrompt(
        """Never forget you are a very professional debater and I am your opponent. You can make full use of your professional debating skills to debate with me. Never flip roles!
        We will debate on a topic a specific person is facing: {topic}
        Remember you are on the this side of the debate: {position}, while I am on the opposite side. NEVER forget the topic and change your stance! Following '===' are the background information and the person's preference of the topic you can refer to. Always remember you are arguing you stance based on the person's needs.
        ===
        Background: {background}
        Preference: {summary}
        ===
        In order to make your justification more convincing, Your thinking pattern should be as follows:
        1. Clearly and concisely state your main claim against my argument.
        2. Explain why you disagree with me. Identify any logical fallacies or vulnerabilities of my last reply. Present your rebuttal, providing evidence. Or, state that even though my argument makes sense, your stance also matters in that case.
        3. Present the reason to support you claim in step 1, which is an in-depth elaboration of your argument. Offer specific evidence, facts, statistics, research findings, expert opinions, examples, historical data, or logical reasoning. You should drill down to just one reason, not outlining multiple reasons. NEVER repeat the argument and reasons you have replied before.
        4. Anticipate and Address Counterarguments. Refute or weaken those counterarguments by providing counter-evidence, highlighting flaws in reasoning, or demonstrating why your position is more valid.
        5. Briefly summarize the main points of the argument and restate the claim.
        Then structure your reply.
        You need to admit defeat if you cannot perform the instruction due to physical, moral, legal reasons or your capability and explain the reasons. In this case, reply with a single word <DEBATE_TOPIC_DONE>.
        Otherwise, your every reply MUST always strictly follow the format below, one argument and one explanation:
        ---
            ###Argument: <YOUR_ARGUMENT> (main idea of your argument in step 1)
            ###Explanation: <YOUR_EXPLANATION> (detailed analysis following the step 2, 3, 4, 5, wrapped all in one paragraph)
        ---""")

    TOPIC_BREAK_DOWN_PROMPT = TextPrompt(
        """Given a two-choice debate topic, list its pro and con side. Don't give your justifications.

        For example:
        TOPIC: I want to find a job as a software engineer but I am considering which city is better, Beijing or Chengdu?
        PRO: Beijing
        CON: Chengdu

        TOPIC: Should I pursue a graduate degree or start my career right now
        PRO: Pursue a graduate degree
        CON: Start a career right now

        Now, your turn. Please break down the following topic into pro and con sides.
        TOPIC: {topic}
        PRO:
        CON:""")

    TOPIC_ABBREVIATE_PROMPT = TextPrompt(
        """Can you abbreviate the topic in no more than 100 characters? Reply with only the topic. Don't add anything else.
        {topic}""")

    TOPIC_SPECIFY_PROMPT = TextPrompt("""I am facing a topic now: {topic}
        Could you specify 5 most important aspects of information are needed to make a decision on this choice? You must only reply a list of most important aspects without any clarification.
        Here is a list of aspects:""")

    TOPIC_COLLECT_BG_PROMPT = TextPrompt(
        """Now your second job is to collect enough information.
        First, collect some objective statistics and data which are important to make a decision from those aspects based on the latest available information.
        Only reply to my two short and concise paragraphs describing your findings, one for each option."""
    )

    TOPIC_COLLECT_INFO_PROMPT = TextPrompt(
        """Then, you can ask me a series of subjective questions in these aspects mentioned earlier to know more about my preferences. When you ask me a question, you must start with <QUESTION> and reply with one question per time.
        Once you have enough information covering alomost most of aspects, you must start with <SUMMARY> and reply with only a summary of information you have collected. Don't give your own suggestions or opinions.
        This summary will be sent to two debaters discussing this question, so please be informative. Now you can ask me the first question."""
    )

    TOPIC_REPHRASE_SUM_PROMPT = TextPrompt(
        """Rephrase your last reply into the third person singular perspective without <SUMMARY>.
        Example: The person (he/she) ...""")

    HOST_PROMPT = TextPrompt(
        """You are a professional judge in a debate. Two debaters are discussing on one topic a specific person he or she is facing: {topic}
        Here is the person's personal preference: {summary}
        Now you will hear their arguments and justifications and you will decide when to end the debate. Your judgement must be only based on the person's personal preferences.
        Each time I will give you one round of debate between them.
        Reply to me '<<<END>>>' if these conditions are met:
            i. Listen to as many rounds of debate as possible.
            ii. Each debater has talked about all of these aspects of information so far: {aspects}.
            iii. One debater's argument and justification is more convincing than the other's in terms of every aspect of the subject's preferences.
        Or reply to me '<<<CONTINUE>>>', then I will give you another round of interaction between them.
        So, your reply should be:
           <<<END>>> Your reasoning for your judgement.
        or
           <<<CONTINUE>>>""")

    HOST_JUDGE_PROMPT = TextPrompt(
        """Now give me your reasoning for your judgement.""")

    HOST_DECISION_PROMPT = TextPrompt(
        """Now you can help me make a decision based on aspects of information you mentioned earlier. And give me reasoning for every aspect. Be concise in your reply."""
    )

    LANGCHAIN_SYS_MSG_PROMPT = """You are a very professional debater and I am your opponent. You can make full use of your professional debating skills to debate with me.
    You have access to the following tools:
    {tool_names}

    We will debate on a topic a specific person is facing: {topic}
    Your stance is: {stance}. Your goal is to persuade your conversation partner of your point of view.
    Following '===' are the background information, the person's preference of the topic. Always remember you are arguing you stance based on the person's preferences.
    ===
    Background:
        {background}
    Preference:
        {summary}
    ===
    I will let you to give your first statement of the debate, or I will give my statement of my stance every time and let you refute me.

    DO argue for your stance and look up information using the tools you have.
    DO cite your sources. DO NOT fabricate fake citations. DO NOT cite any source that you did not look up.
    DO NOT repeat your statements you have made before.

    Use the following format:
    ---
    Start: my instruction or my statement
    Thought: you should always think about what to do next, like what information is needed to support your statement
    Action: the action to take, should be one of [{tool_names}]
    Action Input: the input to the action
    Observation: the result of the action
    ... (this Thought/Action/Action Input/Observation can repeat N times, N equals to 0 or greater than 0)
    Thought: I now collected enough information to make a new reply
    Final Answer: (the final reply against me)
###Argument: your main argument
###Explanation: your explanation of the argument, one paragraph
    ---"""

    LANGCHAIN_HMN_MSG_PROMPT = """Start: {input}
    {agent_scratchpad}"""

    def __init__(self, *args: Any, **kwargs: Any) -> None:
        super().__init__(*args, **kwargs)
        self.update({
            RoleType.DEBATER: self.DEBATER_PROMPT,
            RoleType.HOST: self.HOST_PROMPT,
            "topic_specify": self.TOPIC_SPECIFY_PROMPT,
            "topic_collect_bg": self.TOPIC_COLLECT_BG_PROMPT,
            "topic_collect_info": self.TOPIC_COLLECT_INFO_PROMPT,
            "topic_rephrase_sum": self.TOPIC_REPHRASE_SUM_PROMPT,
            "host_judge": self.HOST_JUDGE_PROMPT,
            "host_decision": self.HOST_DECISION_PROMPT,
            "topic_break_down": self.TOPIC_BREAK_DOWN_PROMPT,
            "topic_abbreviate": self.TOPIC_ABBREVIATE_PROMPT,
            "langchain_sys_msg": self.LANGCHAIN_SYS_MSG_PROMPT,
            "langchain_hmn_msg": self.LANGCHAIN_HMN_MSG_PROMPT,
        })
